Squashed 'include/shark/LinAlg/BLAS/' changes from 52e2a131..b8a85a26
b8a85a26 whoops
64d44230 fixinf warnings in visuals tudio 2017

git-subtree-dir: include/shark/LinAlg/BLAS
git-subtree-split: b8a85a264cd529a7dcfd16421aad60f5bdec9158

+++ kernels/cblas/dense_gemm.hpp
@@ -141,12 +141,12 @@ void dense_gemm(
 	auto storageC = C().raw_storage();
 	dense_gemm(stor_ord, transA, transB, (int)m, (int)n, (int)k, alpha,
 		storageA.values,
-		storageA.leading_dimension,
+		(int)storageA.leading_dimension,
 		storageB.values,
-		storageB.leading_dimension,
+		(int)storageB.leading_dimension,
 		typename MatC::value_type(1),
 		storageC.values,
-		storageC.leading_dimension
+		(int)storageC.leading_dimension
 	);
 }
